refactor: move filters logic to search bar
